client_tests/requirements.txt
# DriftDB Test Suite Dependencies

# Core testing framework
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-timeout>=2.1.0
pytest-xdist>=3.3.0  # Parallel test execution
pytest-cov>=4.1.0    # Coverage reporting

# Database connectivity
psycopg2-binary>=2.9.6

# Test utilities
faker>=19.2.0        # Test data generation
hypothesis>=6.82.0   # Property-based testing

# Performance testing
pytest-benchmark>=4.0.0
memory-profiler>=0.61.0

# Reporting
pytest-html>=3.2.0
pytest-json-report>=1.5.0

# Utilities
colorama>=0.4.6     # Colored output
tabulate>=0.9.0     # Table formatting
pyyaml>=6.0         # Config files
python-dotenv>=1.0.0 # Environment variables